requirements.txt
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
Django==3.0.8
django-cors-headers==3.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
factory-boy==3.2.0
Faker==8.8.0
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.0.1
isort==5.10.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.782
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rest-condition==1.0.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
